Cleanup ACK add try/except for detaching policy (#577)
**Description of your changes:**
Cleanup may fail in edge-case where the policy was already
deleted/detached before the clean-up was run. In this case the role will
never be deleted. Then in future runs of ACK creation in that same
account it would not create the role again, thus the role would have a
missing policy and be unable to run SM jobs

**Testing:**
- [ ] Unit tests pass
- [ ] e2e tests pass
- Details about new tests (If this PR adds a new feature)
- Details about any manual tests performed

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.