test: cherry-pick Introduce mechanism to patch corrupt installations in test clusters (#814)
cherry-pick Introduce mechanism to patch corrupt installations in test
clusters.

Kserve crds sometimes have deletion issues because of a faulty crd. This
is a unique problem to 1.7.

**Which issue is resolved by this Pull Request:**
Resolves #

**Description of your changes:**
- Introduce a deletion mechanism in the installation fixture.

**Testing:**
- [ ] Unit tests pass
- [ ] e2e tests pass - TBD will test with canary.
- Details about new tests (If this PR adds a new feature)
- Details about any manual tests performed

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.

**Which issue is resolved by this Pull Request:**
Resolves #

**Description of your changes:**


**Testing:**
- [ ] Unit tests pass
- [ ] e2e tests pass
- Details about new tests (If this PR adds a new feature)
- Details about any manual tests performed

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.